管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -18,19 +18,19 @@ import views.EmployeeView;
  *
  */
 public class EmployeeAction extends ActionBase {
+
     private EmployeeService service;
 
-    /*
-     * メソッド実行する
+    /**
+     * メソッドを実行する
      */
     @Override
-    public void process() throws ServletException,IOException{
+    public void process() throws ServletException, IOException {
 
-        service=new EmployeeService();
+        service = new EmployeeService();
 
         //メソッドを実行
         invoke();
-
         service.close();
     }
 
@@ -39,47 +39,61 @@ public class EmployeeAction extends ActionBase {
      * @throws ServletException
      * @throws IOException
      */
-    public void index () throws ServletException,IOException{
-        //指定されたページ数の一覧画面に表示されるデータを取得
-        int page=getPage();
-        List<EmployeeView> employees=service.getPerPage(page);
+    public void index() throws ServletException, IOException {
 
-        //すべての従業員データ件数を取得
-        long employeeCount = service.countAll();
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
 
-        putRequestScope(AttributeConst.EMPLOYEES,employees);  //取得した従業員データ
-        putRequestScope(AttributeConst.EMP_COUNT,employeeCount);  //全ての従業員データの件数
-        putRequestScope(AttributeConst.PAGE,page);  //ページ数
-        putRequestScope(AttributeConst.MAX_ROW,JpaConst.ROW_PER_PAGE); //1ページに対するレコードの数
+            //指定されたページ数の一覧画面に表示するデータを取得
+            int page = getPage();
+            List<EmployeeView> employees = service.getPerPage(page);
 
-      //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
-        String flush=getSessionScope(AttributeConst.FLUSH);
-        if(flush!=null) {
-            putRequestScope(AttributeConst.FLUSH,flush);
-            removeSessionScope(AttributeConst.FLUSH);
-        }
+            //全ての従業員データの件数を取得
+            long employeeCount = service.countAll();
 
-        //一覧画面を表示
-        forward(ForwardConst.FW_EMP_INDEX);
+            putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
+            putRequestScope(AttributeConst.PAGE, page); //ページ数
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
+
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
+
+        } //追記
 
     }
-
     /**
      * 新規登録画面を表示する
      * @throws ServletException
      * @throws IOException
      */
-    public void entryNew() throws ServletException,IOException{
-        putRequestScope(AttributeConst.TOKEN,getTokenId()); //CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());//からの従業員インスタンス
+    public void entryNew() throws ServletException, IOException {
 
-        //新規登録画面を表示
-        forward(ForwardConst.FW_EMP_NEW);
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
+
+            //新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        } //追記
     }
+    /**
+     * 新規登録を行う
+     * @throws ServletException
+     * @throws IOException
+     */
+    public void create() throws ServletException, IOException {
 
-    public void create() throws ServletException,IOException{
-        //CSRF対策tokenチェック
-        if(checkToken()) {
+        //CSRF対策 tokenのチェック
+        if (checkAdmin() && checkToken()) { //追記
 
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
@@ -91,54 +105,62 @@ public class EmployeeAction extends ActionBase {
                     null,
                     null,
                     AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
-            //アプリケーションスコープからpepper文字列を取得
-            String pepper =getContextScope(PropertyConst.PEPPER);
 
-            //従業員情報登録
-            List<String> errors=service.create(ev,pepper);
+                    //アプリケーションスコープからpepper文字列を取得
+                    String pepper = getContextScope(PropertyConst.PEPPER);
 
-            if(errors.size()>0) {
-                //登録中にエラーがあった場合
+                    //従業員情報登録
+                    List<String> errors = service.create(ev, pepper);
 
-                putRequestScope(AttributeConst.TOKEN,getTokenId()); //CSRF対策トークン
-                putRequestScope(AttributeConst.EMPLOYEE,ev); //入力された従業員情報
-                putRequestScope(AttributeConst.ERR,errors); //エラーのリスト
+                    if (errors.size() > 0) {
+                        //登録中にエラーがあった場合
 
-                //新規登録画面を再表示
-                forward(ForwardConst.FW_EMP_NEW);
-            }else {
-                //登録中にエラーがなかった場合
+                        putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+                        putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
+                        putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
 
-                //セッションに登録完了のフラッシュメッセージを表示
-                putSessionScope(AttributeConst.FLUSH,MessageConst.I_REGISTERED.getMessage());
+                        //新規登録画面を再表示
+                        forward(ForwardConst.FW_EMP_NEW);
 
-                //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP,ForwardConst.CMD_INDEX);
+                    } else {
+                        //登録中にエラーがなかった場合
+
+                        //セッションに登録完了のフラッシュメッセージを設定
+                        putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
+
+                        //一覧画面にリダイレクト
+                        redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+                    }
+
+                }
             }
-
-        }
-    }
     /**
      * 詳細画面を表示する
      * @throws ServletException
      * @throws IOException
      */
-    public void show() throws ServletException,IOException{
-        //idを条件に従業員データを取得する
-        EmployeeView ev=service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+    public void show() throws ServletException, IOException {
 
-        if(ev==null || ev.getDeleteFlag()==AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
 
-            //データが取得できなかった、または、論理削除されている場合はエラーを表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
-        }
-        putRequestScope(AttributeConst.EMPLOYEE,ev);  //取得した従業員情報
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-        //詳細画面を表示
-        forward(ForwardConst.FW_EMP_SHOW);
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //詳細画面を表示
+            forward(ForwardConst.FW_EMP_SHOW);
+        } //追記
 
     }
+
     /**
      * 編集画面を表示する
      * @throws ServletException
@@ -146,24 +168,27 @@ public class EmployeeAction extends ActionBase {
      */
     public void edit() throws ServletException, IOException {
 
-        //idを条件に従業員データを取得する
-        EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
 
-        if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-            //データが取得できなかった、または論理削除されている場合はエラー画面を表示
-            forward(ForwardConst.FW_ERR_UNKNOWN);
-            return;
-        }
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
 
-        putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
 
-        //編集画面を表示する
-        forward(ForwardConst.FW_EMP_EDIT);
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
 
+            //編集画面を表示する
+            forward(ForwardConst.FW_EMP_EDIT);
+
+        } //追記
     }
-
     /**
      * 更新を行う
      * @throws ServletException
@@ -172,7 +197,7 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) { //追記
             //パラメータの値を元に従業員情報のインスタンスを作成する
             EmployeeView ev = new EmployeeView(
                     toNumber(getRequestParam(AttributeConst.EMP_ID)),
@@ -183,6 +208,7 @@ public class EmployeeAction extends ActionBase {
                     null,
                     null,
                     AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
+
             //アプリケーションスコープからpepper文字列を取得
             String pepper = getContextScope(PropertyConst.PEPPER);
 
@@ -200,6 +226,7 @@ public class EmployeeAction extends ActionBase {
                 forward(ForwardConst.FW_EMP_EDIT);
             } else {
                 //更新中にエラーがなかった場合
+
                 //セッションに更新完了のフラッシュメッセージを設定
                 putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
 
@@ -208,26 +235,48 @@ public class EmployeeAction extends ActionBase {
             }
         }
     }
-        /**
-         * 論理削除を行う
-         * @throws ServletException
-         * @throws IOException
-         */
-        public void destroy() throws ServletException, IOException {
 
-            //CSRF対策 tokenのチェック
-            if (checkToken()) {
+    /**
+     * 論理削除を行う
+     * @throws ServletException
+     * @throws IOException
+     */
+    public void destroy() throws ServletException, IOException {
 
-                //idを条件に従業員データを論理削除する
-                service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+        //CSRF対策 tokenのチェック
+        if (checkAdmin() && checkToken()) { //追記
 
-                //セッションに削除完了のフラッシュメッセージを設定
-                putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+            //idを条件に従業員データを論理削除する
+            service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-                //一覧画面にリダイレクト
-                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
-            }
+            //セッションに削除完了のフラッシュメッセージを設定
+            putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+
+            //一覧画面にリダイレクト
+            redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
+    }
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
+
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
+        } else {
+
+            return true;
         }
 
-
     }
+
+}
